Modified routes.php

+++ src/Controller/AuthorsController.php
@@ -1,10 +1,10 @@
 <?php
 declare(strict_types=1);
 
-// In src/Controller/AuthorsController.php
 namespace App\Controller;
 
 use App\Controller\AppController;
+use Cake\Http\Exception\NotFoundException;
 
 class AuthorsController extends AppController
 {
@@ -12,12 +12,17 @@ class AuthorsController extends AppController
     {
         $authors = $this->Authors->find('all');
         $this->set(compact('authors'));
+        $this->viewBuilder()->setOption('serialize', ['authors']);
     }
 
     public function view($id = null)
     {
         $author = $this->Authors->get($id);
+        if (!$author) {
+            throw new NotFoundException(__('Author not found'));
+        }
         $this->set(compact('author'));
+        $this->viewBuilder()->setOption('serialize', ['author']);
     }
 
     public function add()
@@ -26,37 +31,43 @@ class AuthorsController extends AppController
         if ($this->request->is('post')) {
             $author = $this->Authors->patchEntity($author, $this->request->getData());
             if ($this->Authors->save($author)) {
-                $this->Flash->success(__('The author has been saved.'));
-                return $this->redirect(['action' => 'index']);
+                $this->set(compact('author'));
+                $this->viewBuilder()->setOption('serialize', ['author']);
+                $this->response = $this->response->withStatus(201); // HTTP Created
+            } else {
+                $this->response = $this->response->withStatus(400); // Bad Request
             }
-            $this->Flash->error(__('Unable to add the author.'));
         }
-        $this->set('author', $author);
     }
 
     public function edit($id = null)
     {
         $author = $this->Authors->get($id);
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $author = $this->Authors->patchEntity($author, $this->request->getData());
-            if ($this->Authors->save($author)) {
-                $this->Flash->success(__('The author has been updated.'));
-                return $this->redirect(['action' => 'index']);
-            }
-            $this->Flash->error(__('Unable to update the author.'));
+        if (!$author) {
+            throw new NotFoundException(__('Author not found'));
         }
-        $this->set(compact('author'));
+        if ($this->request->is(['patch', 'post', 'put'])) {
+            $this->Authors->patchEntity($author, $this->request->getData());
+            if ($this->Authors->save($author)) {
+                $this->set(compact('author'));
+                $this->viewBuilder()->setOption('serialize', ['author']);
+            } else {
+                $this->response = $this->response->withStatus(400); // Bad Request
+            }
+        }
     }
 
     public function delete($id = null)
     {
         $this->request->allowMethod(['post', 'delete']);
         $author = $this->Authors->get($id);
-        if ($this->Authors->delete($author)) {
-            $this->Flash->success(__('The author has been deleted.'));
-        } else {
-            $this->Flash->error(__('Unable to delete the author.'));
+        if (!$author) {
+            throw new NotFoundException(__('Author not found'));
         }
-        return $this->redirect(['action' => 'index']);
+        if ($this->Authors->delete($author)) {
+            $this->response = $this->response->withStatus(204); // No Content
+        } else {
+            $this->response = $this->response->withStatus(400); // Bad Request
+        }
     }
 }
